requirements.txt
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
idna==2.8
jsonschema==2.6.0
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.1
Pillow==5.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.22.0
scipy==1.3.1
six==1.11.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.19.9
Unidecode==1.1.1
urllib3==1.25.3
